back/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
langchain==0.1.1
langchain-community==0.0.13
langchain-openai==0.0.2.post1
python-dotenv==1.0.1
firebase-admin==6.5.0